refactor in booking and ticket class to move methods according to ticket to correct class
