+++ src/main/java/pl/agh/edu/mwo/analiza/Booking.java
@@ -10,6 +10,7 @@ import java.util.Random;
 
 import static pl.agh.edu.mwo.analiza.Cinema.removeBookingIfTicketsAreEmpty;
 import static pl.agh.edu.mwo.analiza.Cinema.saveBookingInCinemaStorage;
+import static pl.agh.edu.mwo.analiza.Ticket.creatingTicketForBooking;
 
 @Slf4j
 public class Booking {
@@ -132,51 +133,7 @@ public class Booking {
                         seatsForChildren.size()).multiply(filmDetails.getPriceForChildren()));
     }
 
-    private static List<Ticket> creatingTicketForBooking(List<Seat> seatsForChildren, List<Seat> seatsForAdults, FilmDetails filmDetails, Customer customer) {
-        List<Ticket> tickets = new ArrayList<>();
-        if (customer.hasAccount()) {
-            createTicketForAllSeatsForCustomerWithAccount(seatsForChildren, seatsForAdults, filmDetails, customer, tickets);
-        } else {
-            createTicketForAllSeatsForCustomerWithoutAccount(seatsForChildren, seatsForAdults, filmDetails, tickets);
-        }
-        return tickets;
-    }
 
-    private static void createTickets(List<Seat> seatsList, FilmDetails filmDetails, Customer customer, List<Ticket> tickets) {
-        seatsList.forEach(seat -> {
-            if (seat.isAvailable()) {
-                Ticket ticket = new Ticket(seat.getName(), filmDetails.getFilm().title(), filmDetails.getStartTime(), filmDetails.getStartTime(), customer.getEmail());
-                tickets.add(ticket);
-                seat.lockSeat();
-            } else {
-                System.out.println("Seat: " + seat.getName() + " is already locked. Please choose another seat.");
-            }
-        });
-    }
-
-    private static void createTicketForAllSeatsForCustomerWithAccount(List<Seat> seatsForChildren, List<Seat> seatsForAdults, FilmDetails filmDetails, Customer customer, List<Ticket> tickets) {
-        createTickets(seatsForAdults, filmDetails, customer, tickets);
-        createTickets(seatsForChildren, filmDetails, customer, tickets);
-        log.info("Created tickets for new Booking");
-    }
-
-
-    private static void createTicketForAllSeatsForCustomerWithoutAccount(List<Seat> seatsForChildren, List<Seat> seatsForAdults, FilmDetails filmDetails, List<Ticket> tickets) {
-        createTicketsForNotSpecifiedCustomer(seatsForAdults, filmDetails, tickets);
-        createTicketsForNotSpecifiedCustomer(seatsForChildren, filmDetails, tickets);
-    }
-
-    private static void createTicketsForNotSpecifiedCustomer(List<Seat> seatsList, FilmDetails filmDetails, List<Ticket> tickets) {
-        seatsList.forEach(seat -> {
-            if (seat.isAvailable()) {
-                Ticket ticket = new Ticket(seat.getName(), filmDetails.getFilm().title(), filmDetails.getStartTime(), filmDetails.getStartTime());
-                tickets.add(ticket);
-                seat.lockSeat();
-            } else {
-                System.out.println("Seat: " + seat.getName() + " is already locked. Please choose another seat.");
-            }
-        });
-    }
 
     private void getInfoToPayAtCheckout() {
         System.out.println("Successfully booking process for booking number: " + bookingNumber +
@@ -208,7 +165,6 @@ public class Booking {
                 randomInt(10, 100)
         );
     }
-
     @SuppressWarnings("SameParameterValue")
     private static int randomInt(int min, int max) {
         return new Random().nextInt(max - min) + min;
