--- src/main/java/pl/agh/edu/mwo/analiza/FilmAlreadyExistException.java
@@ -1,8 +1,0 @@
-package pl.agh.edu.mwo.analiza;
-
-public class FilmAlreadyExistException extends RuntimeException {
-
-    public FilmAlreadyExistException(String message) {
-        super(message);
-    }
-}
